requirements.txt
numpy==1.24.3
pandas==1.5.3
Pillow==11.2.1
scikit_learn==1.2.2
Shapely==2.1.0
